refactor: move SSHNP class from bin/sshnp.dart to lib/sshnp.dart refactor: move generation of sessionId into SSHNP.fromCommandLineArgs; add sessionId as final instance variable to SSHNP class, and inject it via SSHNP constructor refactor: combine home_directory.dart, check_non_ascii.dart and check_file_exists.dart into single new file, sshnp_utils.dart refactor: move getUserName() from bin/sshnp.dart to sshnp_utils.dart
