--- lib/check_file_exists.dart
@@ -1,7 +1,0 @@
-import 'dart:async';
-import 'dart:io';
-
-Future<bool> fileExists(String file) async {
-  bool f = await File(file).exists();
-  return f;
-}
