+++ lib/sshnp_utils.dart
@@ -1,0 +1,49 @@
+import 'dart:io';
+
+/// Get the home directory or null if unknown.
+String? getHomeDirectory() {
+  switch (Platform.operatingSystem) {
+    case 'linux':
+    case 'macos':
+      return Platform.environment['HOME'];
+    case 'windows':
+      return Platform.environment['USERPROFILE'];
+    case 'android':
+      // Probably want internal storage.
+      return '/storage/sdcard0';
+    case 'ios':
+      // iOS doesn't really have a home directory.
+      return null;
+    case 'fuchsia':
+      // I have no idea.
+      return null;
+    default:
+      return null;
+  }
+}
+
+/// Get the local username or null if unknown
+String? getUserName() {
+  Map<String, String> envVars = Platform.environment;
+  if (Platform.isLinux || Platform.isMacOS) {
+    return envVars['USER'];
+  } else if (Platform.isWindows) {
+    return envVars['USERPROFILE'];
+  }
+  return null;
+}
+
+Future<bool> fileExists(String file) async {
+  bool f = await File(file).exists();
+  return f;
+}
+
+bool checkNonAscii(String test) {
+  var extra = test.replaceAll(RegExp(r'[a-zA-Z0-9_]*'), '');
+  if ((extra != '') || (test.length > 15)) {
+    return true;
+  } else {
+    return false;
+  }
+}
+
