+++ lib/version.dart
@@ -1,6 +1,6 @@
 import 'dart:io';
 
-//Print version number
+/// Print version number
 void version() {
   final String version = "3.3.0";
   stdout.writeln('Version : $version');
